fix: Handle unknown errors in main function
This change updates the error handling in the main function to catch unknown errors. Previously, only errors that were instances of the `Error` class were caught. Now, any error that is thrown will be caught and logged.

This change ensures that the script will not crash due to an unhandled error.